refactor: export core-api types (#1948)
