+++ packages/core-api/__tests__/repositories/transactions.test.ts
@@ -3,7 +3,8 @@ import "jest-extended";
 
 import { crypto } from "@arkecosystem/crypto";
 import genesisBlock from "../../../core-test-utils/src/config/testnet/genesisBlock.json";
-import { TransactionsRepository } from "../../src/repositories/transactions";
+// noinspection TypeScriptPreferShortImport
+import { TransactionsRepository } from "../../dist/repositories/transactions";
 import { setUp, tearDown } from "../__support__/setup";
 
 let repository;
