+++ packages/core-api/package.json
@@ -7,7 +7,8 @@
         "Brian Faust <brian@ark.io>"
     ],
     "license": "MIT",
-    "main": "dist/index.js",
+    "main": "dist/index",
+    "types": "dist/index",
     "files": [
         "dist"
     ],
