+++ packages/core-api/src/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from "./repository";
