+++ packages/core-api/src/repositories/index.ts
@@ -1,5 +1,12 @@
 import { BlockRepository } from "./blocks";
 import { TransactionsRepository } from "./transactions";
 
-export const blocksRepository = new BlockRepository();
-export const transactionsRepository = new TransactionsRepository();
+const blocksRepository = new BlockRepository();
+const transactionsRepository = new TransactionsRepository();
+
+export {
+    blocksRepository,
+    transactionsRepository,
+    BlockRepository,
+    TransactionsRepository
+}
