+++ packages/core-api/src/repositories/repository.ts
@@ -2,7 +2,7 @@ import { app } from "@arkecosystem/core-container";
 import { PostgresConnection } from "@arkecosystem/core-database-postgres";
 import { TransactionPool } from "@arkecosystem/core-interfaces";
 import snakeCase from "lodash/snakeCase";
-import { IRepository } from "../interfaces/repository";
+import { IRepository } from "../interfaces";
 
 export abstract class Repository implements IRepository {
     public database = app.resolvePlugin<PostgresConnection>("database");
