+++ packages/core-api/src/server.ts
@@ -1,17 +1,15 @@
 import { app } from "@arkecosystem/core-container";
 import { createSecureServer, createServer, mountServer, plugins } from "@arkecosystem/core-http-utils";
+import { Logger } from "@arkecosystem/core-interfaces";
 import Hapi from "hapi";
 
 export class Server {
-    private config: any;
-    private logger: any;
+    private logger = app.resolvePlugin<Logger.ILogger>("logger");
 
     private http: any;
     private https: any;
 
-    public constructor(config: any) {
-        this.config = config;
-        this.logger = app.resolvePlugin("logger");
+    public constructor(private config: any) {
     }
 
     public async start(): Promise<void> {
